Modify to apply when no transaction but return cash

+++ src/processor/Output/Action/Action12/ReturnCashImpl2.java
@@ -10,7 +10,7 @@ public class ReturnCashImpl2 extends ReturnCash {
     @Override
     public void returnCash() {
         DataStore2 ds2 = (DataStore2) this.ds;
-        float rCash = (ds2.cash > 0) ? ds2.cash : ds2.temp_cash;
+        int rCash = (ds2.cash > 0) ? ds2.cash : ds2.temp_cash;
         float remain = rCash - ds2.total;
         System.out.println("[System] Remainder has been returned. Please remember to take your cash!");
         System.out.printf("[Info] Cash returned: $ %f\n", remain);
